uhh.. read desc
I'm pushing the most current rot in UW but plans have changed and the rot will be more centered around UW_I01 and spread outward accordingly, try to paint what would've happened as a result of rot getting through GATE_UW_SS if you have the motivation to. if no one takes the oppurtunity for themselves  I'll do it myself.